% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackScholes.R
\name{analyticBlackScholesPrice}
\alias{analyticBlackScholesPrice}
\title{Compute European option prices via analytic formulas for Poisson driven prices.}
\usage{
analyticBlackScholesPrice(strike, maturity, type, spot, rate, div, volat)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{div}{the continuous dividend yield rate}

\item{volat}{the volatility level}
}
\value{
numeric
}
\description{
{European option pricing via exact formulas under Levy-Poisson model dynamics for
vanilla options like calls and puts.}
}
