% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcpp.R
\name{analyticPoissonPrice}
\alias{analyticPoissonPrice}
\title{Compute European option prices via analytic formulas for Poisson driven prices.}
\usage{
analyticPoissonPrice(strike, maturity, type, spot, rate, a, b)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{a}{the coefficient of the Poisson process}

\item{b}{the drift-compensation coefficient}
}
\value{
numeric
}
\description{
{European option pricing via exact formulas under Levy-Poisson model dynamics for
vanilla options like calls and puts.}
}
