% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpoisPDE}
\alias{cpoisPDE}
\title{Price options via Compensated Poisson PDE}
\usage{
cpoisPDE(strike, maturity, type, spot, r, a, b, res, american = TRUE)
}
\arguments{
\item{strike}{the strike price of the option contract}

\item{maturity}{the time until maturity}

\item{type}{either "put" or "call}

\item{spot}{the spot price}

\item{r}{the risk free rate}

\item{a}{coefficient of the Poisson process}

\item{b}{coefficient of the compensating drift}

\item{res}{the grid resolution}

\item{american}{bool for american options}
}
\value{
numeric
}
\description{
{Compute European/American options via a finite difference solver for the
PDE for compensated Poisson log-returns.}
}
