% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedBackouts.R
\name{impliedRate}
\alias{impliedRate}
\title{Back out implied rate via Black-Scholes PDE}
\usage{
impliedRate(fee, strike, maturity, type, spot, volat)
}
\arguments{
\item{fee}{the given market price of the option}

\item{strike}{the given strike price of the option}

\item{maturity}{the maturity of the option in trading years until expiration}

\item{type}{the type of option: "call" or "put"}

\item{spot}{the spot price of the underlying}

\item{volat}{the volatility level}
}
\value{
numeric
}
\description{
{Call to base r \code{uniroot} to backout the implied rate of an option given
 all other parameters of the Black-Scholes model.}
}
