% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedBackouts.R
\name{impliedVolatility}
\alias{impliedVolatility}
\title{Back out implied volatility via Black-Scholes PDE}
\usage{
impliedVolatility(fee, strike, maturity, type, spot, rate)
}
\arguments{
\item{fee}{the given market price of the option}

\item{strike}{the given strike price of the option}

\item{maturity}{the maturity of the option in trading years until expiration}

\item{type}{the type of option: "call" or "put"}

\item{spot}{the spot price of the underlying}

\item{rate}{the risk-free rate}
}
\value{
numeric
}
\description{
{Call to base r \code{uniroot} to backout the implied volatility of an option given
 all other parameters of the Black-Scholes model.}
}
