% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackScholes.R
\name{monteCarloBlackScholes}
\alias{monteCarloBlackScholes}
\title{Compute European option prices via Monte-Carlo means under Black-Scholes dynamics}
\usage{
monteCarloBlackScholes(
  strike,
  maturity,
  type,
  spot,
  rate,
  div,
  volat,
  n = 10^6
)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{div}{the (continuous) dividend yield rate}

\item{volat}{the volatility level}

\item{n}{number of variates to use in Monte-Carlo mean}
}
\value{
numeric
}
\description{
{European option pricing via Monte-Carlo means under Black-Scholes dynamics for
vanilla options like calls and puts.}
}
