% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kou.R
\name{monteCarloKouPrice}
\alias{monteCarloKouPrice}
\title{Comptue a conditional expectation via Monte-Carlo simulation/averaging}
\usage{
monteCarloKouPrice(
  strike,
  maturity,
  spot,
  parameters,
  type = "call",
  n = 20000
)
}
\arguments{
\item{strike}{the strike sprice}

\item{maturity}{the maturity of the option contract}

\item{spot}{the spot price}

\item{parameters}{vector of model parameters}

\item{type}{the payoff to use}

\item{n}{number of variates to sample}
}
\description{
{Compute a conditional expectation via discounting and averaging}
}
