% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton.R
\name{monteCarloMertonPrice}
\alias{monteCarloMertonPrice}
\title{Merton's jump-diffusion model analytic price formula for European options}
\usage{
monteCarloMertonPrice(
  strike,
  maturity,
  type,
  spot,
  rate,
  div,
  volat,
  lambda,
  jm,
  jv,
  n = 10^4
)
}
\arguments{
\item{strike}{The strike price}

\item{maturity}{the year until expirations}

\item{type}{What type of option, put or call}

\item{spot}{the spot price}

\item{rate}{the drift rate}

\item{div}{The dividend yield rate}

\item{volat}{the diffusion coefficients of each component (a vector)}

\item{lambda}{the yearly jump-rate}

\item{jm}{the jump mean}

\item{jv}{the jump volatility}

\item{n}{number of variates}
}
\value{
numeric
}
\description{
{Monte-Carlo approximation to European price under Merton's jump diffusion.}
}
