% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcpp.R
\name{monteCarloPoisson}
\alias{monteCarloPoisson}
\title{Compute European option prices via Monte-Carlo means under Levy-Poisson dynamics}
\usage{
monteCarloPoisson(strike, maturity, type, spot, rate, a, b, n = 10^6)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{a}{the coefficient of the Poisson process}

\item{b}{the drift-compensation coefficient}

\item{n}{number of variates to use in Monte-Carlo mean}
}
\value{
numeric
}
\description{
{European option pricing via Monte-Carlo means under Levy-Poisson model dynamics for
vanilla options like calls and puts.}
}
