% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pricerPIDE}
\alias{pricerPIDE}
\title{Price options via Black-Scholes PDE}
\usage{
pricerPIDE(
  strike,
  maturity,
  type,
  spot,
  r,
  q,
  v,
  lambda,
  jumpDistr,
  jumpParam,
  res,
  american = TRUE
)
}
\arguments{
\item{strike}{the strike price of the option contract}

\item{maturity}{the time until maturity}

\item{type}{either "put" or "call"}

\item{spot}{the spot price}

\item{r}{the risk free rate}

\item{q}{the continuous dividend yield}

\item{v}{the volatiltiy}

\item{lambda}{the mean number of jumps annually}

\item{jumpDistr}{name of jump-size distribution: "norm", "kou" or "dkou"}

\item{jumpParam}{vector of jump parameters, see details}

\item{res}{the time-space resolution}

\item{american}{bool for american options}
}
\value{
numeric
}
\description{
{Compute European/American options via a finite difference solver for the
jump diffusion PIDE.}
}
