% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testAnalyticVsMonteCarlo}
\alias{testAnalyticVsMonteCarlo}
\title{Compare accuracies of Monte-Carlo and analytic formulas for option prices}
\usage{
testAnalyticVsMonteCarlo(
  strike,
  maturity,
  type,
  spot,
  rate,
  div,
  volat,
  a,
  b,
  lambda,
  jm,
  jv,
  n = 10^6
)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{div}{the continuous dividend yield rate}

\item{volat}{the volatility level}

\item{a}{the coefficient of the Poisson process}

\item{b}{the drift-compensation coefficient}

\item{lambda}{the mean number of jumps annually}

\item{jm}{the mean jump-size}

\item{jv}{the standard deviation of jump-size}

\item{n}{number of variates to use in Monte-Carlo mean}
}
\value{
numeric
}
\description{
{Compare convergence of Monte-Carlo method to analytic formulas for all models.}
}
