% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testDisplacedKouEuro}
\alias{testDisplacedKouEuro}
\title{Test Displaced-Kou pricer for European option prices}
\usage{
testDisplacedKouEuro(
  strike,
  maturity,
  type,
  spot,
  rate,
  div,
  volat,
  lambda,
  p,
  alpha,
  beta,
  ku,
  kd,
  N = 300,
  M = 300
)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{div}{the continuous dividend yield rate}

\item{volat}{the volatility level}

\item{lambda}{the mean number of jumps per year}

\item{p}{the probability of jump upward}

\item{alpha}{the mean jump size up}

\item{beta}{the mean jump size down}

\item{ku}{displacement upward}

\item{kd}{displacement downard}

\item{N}{time resolution}

\item{M}{space resolution}
}
\value{
matrix
}
\description{
{Verify convergence between the three methods: analytic, Monte-Carlo and PDE}
}
