% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testMertonEuro}
\alias{testMertonEuro}
\title{Test Merton pricer for European option prices}
\usage{
testMertonEuro(
  strike,
  maturity,
  type,
  spot,
  rate,
  div,
  volat,
  lambda,
  jm,
  jv,
  N = 300,
  M = 300
)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{div}{the continuous dividend yield rate}

\item{volat}{the volatility level}

\item{lambda}{the mean number of jumps per year}

\item{jm}{the mean jump size}

\item{jv}{the jump size volatility}

\item{N}{time resolution}

\item{M}{space resolution}
}
\value{
matrix
}
\description{
{Verify convergence between the three methods: analytic, Monte-Carlo and PDE}
}
