% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testPoissonEuro}
\alias{testPoissonEuro}
\title{Test Poisson-pricer European option pricers}
\usage{
testPoissonEuro(strike, maturity, type, spot, rate, a, b, N = 300, M = 300)
}
\arguments{
\item{strike}{the strike price of the option}

\item{maturity}{the maturity in trading years until expiration}

\item{type}{type of option, "call" or "put"}

\item{spot}{the current spot price}

\item{rate}{the risk-neutral rate}

\item{a}{the Poisson coefficient}

\item{b}{the drift coefficient}

\item{N}{time resolution}

\item{M}{space resolution}
}
\value{
matrix
}
\description{
{Verify convergence between the three methods: analytic, Monte-Carlo and PDE}
}
