% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kou.R
\name{monteCarloKouPrice}
\alias{monteCarloKouPrice}
\title{Comptue a conditional expectation via Monte-Carlo simulation/averaging}
\usage{
monteCarloKouPrice(
  strike,
  maturity,
  type = "call",
  spot,
  rate,
  div,
  volat,
  lambda,
  parameters,
  n = 20000
)
}
\arguments{
\item{strike}{the strike price}

\item{maturity}{the maturity of the option contract}

\item{type}{the payoff to use}

\item{spot}{the spot price}

\item{rate}{the risk-neutral rate of return}

\item{div}{the continuous dividend yield}

\item{volat}{the volatility level}

\item{lambda}{the mean number of jumps per annum}

\item{parameters}{vector of model parameters}

\item{n}{number of variates to sample}
}
\description{
{Compute a conditional expectation via discounting and averaging}
}
